Make input amount responsive to price changes
By re-writing the reducer updating logic on the buying.inputAmountPair
state and the display logic in the AmountInput component. The input box
in focus has the value frozen while the other one will fluctuate.

Also refactored the input box in Currency Block into AmountInput, a smaller component
on its own